Delete and Fetch fixed with right return codes

+++ capitals/main.py
@@ -91,6 +91,9 @@ def fetch_country(id):
     query = ds.query(kind=kind)
     query.order = ['id']
     result = get_fetch_results(query, id)
+    if len(result) == 0:
+        return make_response("Capital not found", 404)
+
     return jsonify(result)
 
 @app.route('/api/capitals/<int:id>', methods=['DELETE'])
@@ -98,19 +101,26 @@ def delete_country(id):
     client = datastore.Client(project="hackathon-team-016")
     kind = "Countries16"
         
+    if isinstance(id, int) == False:
+        logging.info("input not int")
+        return make_response("Unexpected error", 500)
+
     #Fetch entity with id
     query = client.query(kind=kind)
     query.add_filter('id', "=", id)
     entities = list(query.fetch())
-    entity = entities[0]
-    deleteKey = entity.key
-    
-    logging.info(deleteKey)
-    logging.info(entity)
-    logging.info("Satya - Delete using key")   
-    client.delete(deleteKey)
+    if len(entities) > 0:
+        entity = entities[0]
+        deleteKey = entity.key
+        
+        logging.info(deleteKey)
+        logging.info(entity)
+        logging.info("Satya - Delete using key")   
+        client.delete(deleteKey)
 
-    return "deleted"
+        return "deleted"
+    else:
+        return make_response("Capital record not found", 404)
 
 def get_fetch_results(query, id):
     results = list()
@@ -119,12 +129,12 @@ def get_fetch_results(query, id):
             results.append(dict(entity))
     return results
 
-
 @app.errorhandler(500)
 def server_error(e):
     # Log the error and stacktrace.
     logging.exception('An error occurred during a request.')
-    return 'An internal error occurred.', 500
+    #return 'An internal error occurred.', 500
+    return make_response('Unexpected error', 500)
 
 if __name__ == '__main__':
     # Used for running locally
